Update Assistant model and routes for creating and removing an assistant

+++ src/routes/assistant.routes.ts
@@ -2,7 +2,11 @@ import express from 'express';
 import { getJob, getJobs, rerunJob } from '../services/agenda/agenda.service';
 import { handleUserInput } from '../services/assistant.service';
 import { Assistant } from '../models/Assistant';
-import { updateAssistantById } from '../services/oai.assistant.service';
+import {
+  createAssistant,
+  deleteAssistantById,
+  updateAssistantById,
+} from '../services/oai.assistant.service';
 
 const assistantRouter = express.Router();
 
@@ -23,7 +27,6 @@ assistantRouter.get('/:id', async (req, res) => {
 });
 
 assistantRouter.put('/:id', async (req, res) => {
-
   const { id } = req.params;
   const assistantData = req.body;
 
@@ -39,18 +42,30 @@ assistantRouter.put('/:id', async (req, res) => {
     assistant.llmModel,
     assistant.llmPrompt,
   );
-  
-  res.send(assistant);
 
+  res.send(assistant);
 });
 
 assistantRouter.post('/', async (req, res) => {
   const assistantData = req.body;
   const newAssistant = new Assistant(assistantData);
+
   try {
     await newAssistant.save();
+
+    const openAIAssistant = await createAssistant(
+      assistantData.name,
+      assistantData.description,
+      assistantData.llmModel,
+      assistantData.llmPrompt,
+    );
+
+    newAssistant.assistantId = openAIAssistant.id;
+    await newAssistant.save();
+
     res.send(newAssistant);
   } catch (err) {
+
     if (err instanceof Error && 'code' in err && err.code === 11000) {
       res.status(400).send({
         message:
@@ -58,7 +73,7 @@ assistantRouter.post('/', async (req, res) => {
       });
     } else {
       res.status(500).send({
-        message: 'An error occurred while trying to create the assistant.',
+        message: `An error occurred while trying to create the assistant : ${err}`,
       });
     }
   }
@@ -66,6 +81,18 @@ assistantRouter.post('/', async (req, res) => {
 
 assistantRouter.delete('/:id', async (req, res) => {
   const { id } = req.params;
+  const assistant = await Assistant.findById(id);
+  if (!assistant) {
+    return res.status(404).send({ message: 'Assistant not found' });
+  }
+
+  // Delete the assistant from OpenAI
+  const deleted = await deleteAssistantById(assistant.assistantId);
+  if (!deleted) {
+    return res.status(500).send({ message: 'Failed to delete assistant from OpenAI' });
+  }
+
+  // Delete the assistant from the local database
   await Assistant.findByIdAndDelete(id);
   res.send({ message: 'Assistant deleted successfully' });
 });
